requirements.txt
EasyProcess==0.2.2
Pillow==3.3.0
selenium==2.53.6
